+++ app/src/main/java/com/kevan/hangry/domain/model/DashboardWidget.kt
@@ -12,7 +12,6 @@ enum class WidgetType {
     HEART_METRICS,
     VITALS_CARD,
     SLEEP_SUMMARY,
-    TRAINING_LOAD,
     CALORIE_BURN,
     AI_SHORTCUTS,
     CUSTOM_METRIC
@@ -21,14 +20,11 @@ enum class WidgetType {
 @Serializable
 enum class MetricType {
     STEPS,
-    ACTIVE_MINUTES,
     ACTIVE_CALORIES,
     RHR,
     HRV,
     VO2_MAX,
     SPO2,
-    RESPIRATORY_RATE,
-    BLOOD_PRESSURE,
     SLEEP_DURATION,
     DAY_STRAIN,
     STRESS,
@@ -99,39 +95,32 @@ data class DashboardWidget(
                 order = 5
             ),
             DashboardWidget(
-                id = "training_load",
-                type = WidgetType.TRAINING_LOAD,
-                title = "Cardiovascular Strain & Load",
-                isVisible = true,
-                order = 6
-            ),
-            DashboardWidget(
                 id = "heart_metrics",
                 type = WidgetType.HEART_METRICS,
                 title = "Heart Rate & HRV",
                 isVisible = true,
-                order = 7
+                order = 6
             ),
             DashboardWidget(
                 id = "vitals_card",
                 type = WidgetType.VITALS_CARD,
                 title = "Key Vitals",
                 isVisible = true,
-                order = 8
+                order = 7
             ),
             DashboardWidget(
                 id = "ai_shortcuts",
                 type = WidgetType.AI_SHORTCUTS,
                 title = "Nutrition & Posture",
                 isVisible = true,
-                order = 9
+                order = 8
             ),
             DashboardWidget(
                 id = "calorie_burn",
                 type = WidgetType.CALORIE_BURN,
                 title = "Calorie Expenditure Breakdown",
                 isVisible = false, // disabled by default since DailyActivityRings shows active calories, user can enable
-                order = 10
+                order = 9
             )
         )
     }